requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
Werkzeug==2.0.1
python-dotenv==0.19.0
Flask-Login==0.5.0
bcrypt==3.2.0
Flask-WTF==0.15.1
email-validator==1.1.3
python-dateutil==2.8.2 